requirements.txt
openai~=1.35.13
streamlit~=1.36.0
python-dotenv~=1.0.1